Move unit tests out of org.ensime.test, moved util functions into test/org/ensime/util and placed all other tests in appropriate packages.
